+++ src/compas_timber/model.py
@@ -138,6 +138,16 @@ class TimberModel(Model):
         return candidates
 
     @property
+    def unpromoted_joint_candidates(self) -> set[JointCandidate]:
+        unpromoted = set()
+        for edge in self._graph.edges():
+            edge_candidate = self._graph.edge_attribute(edge, "candidates")
+            edge_joint = self._graph.edge_attribute(edge, "joints")
+            if edge_candidate is not None and edge_joint is None:
+                unpromoted.add(edge_candidate)
+        return unpromoted
+
+    @property
     def topologies(self):
         return self._topologies